% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_variance.R
\name{avar}
\alias{avar}
\title{Asymptotic variance for the kappas.}
\usage{
avar(x, sigma, mu, type, fleiss)
}
\arguments{
\item{x}{Covariance matrix as calculated by \code{cov_mat_kappa}.}

\item{type}{Either \code{adf}, \code{normal}, or \code{elliptical}.}

\item{fleiss}{If \code{TRUE} calculates variance for Fleiss kappa, else
Conger's kappa.}
}
\value{
The asymptotic variance for the kappas.
}
\description{
Asymptotic variance for the kappas.
}
\keyword{internal}
