% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_variance.R
\name{avar_}
\alias{avar_}
\title{Asymptotic variance for the kappas.}
\usage{
avar_(mat, mu, sigma, fleiss)
}
\arguments{
\item{mat}{Covariance matrix as calculated by \code{cov_mat_kappa}.}

\item{mu}{Vector of means.}

\item{sigma}{Covariance matrix}

\item{fleiss}{If \code{TRUE} calculates variance for Fleiss kappa, else
Conger's kappa.}
}
\value{
The asymptotic variance for the kappas.
}
\description{
Asymptotic variance for the kappas.
}
\keyword{internal}
