% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_variance.R
\name{avar_bp}
\alias{avar_bp}
\title{Asymptotic variance for Brennan-Prediger.}
\usage{
avar_bp(x, type, c1)
}
\arguments{
\item{x}{Covariance matrix as calculated by \code{cov_mat_kappa}.}

\item{type}{Either \code{adf}, \code{normal}, or \code{elliptical}.}

\item{c1}{Denominator in BP.}
}
\value{
The asymptotic variance for the kappas.
}
\description{
Asymptotic variance for Brennan-Prediger.
}
\keyword{internal}
