% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bp_get_c1}
\alias{bp_get_c1}
\title{Get the required value of c1.}
\usage{
bp_get_c1(values, kind)
}
\arguments{
\item{values}{Vector of values.}

\item{kind}{The kind of c1 requested.}
}
\value{
The value of c1.
}
\description{
Get the required value of c1.
}
\keyword{internal}
