% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_bootstrap.R, R/long_bootstrap.R, R/misc.R
\name{ci_boot_aggr}
\alias{ci_boot_aggr}
\alias{ci_boot}
\alias{ci}
\alias{ci_asymptotic}
\title{Calculates asymptotic confidence intervals.}
\usage{
ci_boot_aggr(est, sd, calc, transformer, quants, n_reps, est_fun, var_fun)

ci_boot(
  x,
  est,
  sd,
  type,
  transformer,
  quants,
  n_reps,
  fleiss = FALSE,
  kind = NULL,
  values = NULL
)

ci_asymptotic(est, sd, n, transformer, quants)
}
\arguments{
\item{est, sd}{The estimate and estimated standard deviation.}

\item{transformer}{A transformer object.}

\item{quants}{Quantiles for the confidence interval.}

\item{n_reps}{Number of bootstrap samples if \code{bootstrap = TRUE}. Ignored if
\code{bootstrap = FALSE}.}

\item{x}{Data to estimate kappa on.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{fleiss}{If \code{TRUE}, calculates Fleiss' kappa. If not, calculates
Conger's kappa.}

\item{n}{Number of observations.}
}
\description{
Calculates asymptotic confidence intervals.
}
\keyword{internal}
