% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{cov_mat}
\alias{cov_mat}
\title{The covariance matrix of the elements involved in the inference.}
\usage{
cov_mat(p, mu, sigma, gamma, x)
}
\arguments{
\item{p}{Vector of probabilities for being missing.}

\item{mu}{Vector of means.}

\item{sigma}{Covariance matrix}

\item{gamma}{Covariance matrix of the covariances.}
}
\value{
The covariance matrix of x, y, z
}
\description{
Returns the covariance matrix of x, y, z as defined in the paper.
}
\keyword{internal}
