% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fleiss1971.R
\docType{data}
\name{dat.fleiss1971}
\alias{dat.fleiss1971}
\title{Agreement study from Fleiss, J. L. (1971)}
\format{
The tibble contains the five columns "depression",
"personality disorder", "schizophrenia" "neurosis", and "other".
The content of the \code{ij}th cell is the number of raters who diagnosed
the \code{i}th patient as having mental illness \code{j}.
}
\source{
The data was scraped by hand from Fleiss (1971).
}
\usage{
dat.fleiss1971
}
\description{
Fleiss, J. L. (1971) is best known for introducing Fleiss' kappa. In Table 1
he presents diagnosis data of psychiatric patients. The data is on Fleiss
form, with \code{n = 30} patients diagnosed by \code{6} psychiatrists each. It is
not the case that the same set of six psychiatrists diagnosed every
patient.
}
\references{
Fleiss, J. L. (1971). Measuring nominal scale agreement among many raters.
Psychological Bulletin, 76(5), 378–382. https://doi.org/10.1037/h0031619
}
\keyword{datasets}
