% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gwet2014.R
\docType{data}
\name{dat.gwet2014}
\alias{dat.gwet2014}
\title{Agreement study from Gwet (2014)}
\format{
A \code{n} times \code{R} matrix. There are \code{n = 20} row corresponding to the
different items. Each of the \code{R = 5} columns contains the ratings of
the \code{j}th judge.
}
\source{
Gwet, K. L. (2014). Handbook of Inter-Rater Reliability. Advanced Analytics, LLC.
}
\usage{
dat.zapf2016
}
\description{
Agreement study (n = 20) from Gwet (2014), p. 125. This data set contains
missing ratings.
}
\keyword{datasets}
