% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-klein2018.R
\docType{data}
\name{dat.klein2018}
\alias{dat.klein2018}
\title{Agreement study from Klein (2018)}
\format{
A \code{n} times \code{R} matrix. There are \code{n = 20} row corresponding to the
different items. Each of the \code{R = 5} columns contains the ratings of
the \code{j}th judge.
}
\source{
Klein, D. (2018). Implementing a General Framework for Assessing Interrater
Agreement in Stata. The Stata Journal, 18(4), 871–901.
https://doi.org/10.1177/1536867X1801800408
}
\usage{
dat.zapf2016
}
\description{
Agreement study (n = 10) from Klein (2018). This data set contains
missing ratings.
}
\keyword{datasets}
