% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{fleiss_aggr_var}
\alias{fleiss_aggr_var}
\title{Variance for Fleiss' kappa with aggregated raters.}
\usage{
fleiss_aggr_var(x, values = seq(ncol(x)))
}
\arguments{
\item{x}{Data on Fleiss form.}

\item{values}{Values to attach to each column on the Fleiss form data.
Defaults to \code{1:C}, where \code{C} is the number of categories.}
}
\value{
Calculated value of Fleiss' kappa.
}
\description{
Variance for Fleiss' kappa with aggregated raters.
}
