% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fleissci}
\alias{fleissci}
\alias{congerci}
\alias{cohenci}
\title{Confidence intervals for quadratically weighted Fleiss' kappa and
Conger's kappa.}
\usage{
fleissci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

congerci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

cohenci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)
}
\arguments{
\item{x}{Input data data can be converted to a matrix using \code{as.matrix}.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{transform}{One of \code{"none"}, \code{"log"}, \code{"fisher"}, and \verb{"arcsin}.
Defaults to \code{"none"}.}

\item{conf_level}{Confidence level. Defaults to \code{0.95}.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{bootstrap}{If \code{TRUE}, performs a studentized bootstrap with \code{n_reps}
repetitions. Defaults to \code{FALSE}.}

\item{n_reps}{Number of bootstrap samples if \code{bootstrap = TRUE}. Ignored if
\code{bootstrap = FALSE}. Defaults to \code{1000}.}
}
\value{
A vector of class \code{fleissci} containing the confidence end points.
The arguments of the function call are included as attributes.
}
\description{
Confidence intervals for quadratically weighted Fleiss' kappa and
Conger's kappa.
}
