% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gamma_est}
\alias{gamma_est}
\title{Gamma matrix}
\usage{
gamma_est(x, sigma, type = "adf")
}
\arguments{
\item{x}{A numeric matrix of observations.}

\item{sigma}{Covariance matrix of the data.}

\item{type}{One of \code{adf}, \code{normal}, \code{elliptical} or \code{unbiased}.}
}
\value{
The sample estimate of the gamma matrix.
}
\description{
Calculate the gamma matrix from a matrix of observations.
}
\keyword{internal}
