% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gamma_est_unbiased}
\alias{gamma_est_unbiased}
\title{Unbiased asymptotic covariance matrix.}
\usage{
gamma_est_unbiased(x, sigma)
}
\arguments{
\item{x}{Data.}

\item{sigma}{Covariance matrix of the data.}
}
\value{
Unbiased asymptotic covariance matrix.
}
\description{
Unbiased asymptotic covariance matrix.
}
\keyword{internal}
