% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_diag_indices}
\alias{get_diag_indices}
\title{Get the indices of the diagonal for the variances.}
\usage{
get_diag_indices(r)
}
\arguments{
\item{r}{Number of raters.}
}
\value{
A vector of \code{TRUE} where the variances are.
}
\description{
Get the indices of the diagonal for the variances.
}
\keyword{internal}
