% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_kappa.R
\name{pi_mat}
\alias{pi_mat}
\title{Calculate the capital pi matrix.}
\usage{
pi_mat(p, vech = TRUE)
}
\arguments{
\item{p}{Vector of probabilities for being missing.}

\item{vech}{Does not affect anything. The half-vectorized matrix is always
returned.}
}
\value{
The capital pi matrix.
}
\description{
The capital pi matrix is multiplied element-wise with the asymptotic
covariance matrix of s to correct for missing values.
}
\keyword{internal}
