% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pi_mat_empirical}
\alias{pi_mat_empirical}
\title{Calculates the empirical capital pi matrix.}
\usage{
pi_mat_empirical(x)
}
\arguments{
\item{x}{Vector of probabilities for being missing.}
}
\value{
The capital pi matrix.
}
\description{
Calculates the empirical capital pi matrix.
}
\keyword{internal}
