% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadagree-package.R
\docType{package}
\name{quadagree-package}
\alias{quadagree}
\alias{quadagree-package}
\title{quadagree: Inference for the quadratically weighted Fleiss' kappa, Conger's kappa, Cohen's kappa, and Brennan-Prediger coefficient}
\description{
Provides functions for the estimating quadratically weighted agreement coefficients
Fleiss' kappa and Conger's kappa (multirater Cohen's kappa) with inference. It supports
both complete and incomplete data (missing completely at random), and offers
confidence intervals via the delta method or a studentized bootstrap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonasmoss.github.io/quadagree/}
}

}
\author{
\strong{Maintainer}: Jonas Moss \email{jonas.moss.statistics@gmail.com} (\href{https://orcid.org/0000-0002-6876-6964}{ORCID})

}
