% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{quadagree}
\alias{quadagree}
\alias{fleissci}
\alias{fleiss_aggr_ci}
\alias{congerci}
\alias{bp_aggr}
\alias{cohenci}
\title{Confidence intervals for the quadratically weighted Fleiss' kappa and
Conger's (Cohen's) kappa}
\usage{
fleissci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

fleiss_aggr_ci(
  x,
  values = seq(ncol(x)),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

congerci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

bp_aggr(x, values = seq(ncol(x)), kind = 1)

cohenci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)
}
\arguments{
\item{x}{Input data data can be converted to a matrix using \code{as.matrix}.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}. Ignored in \code{fleiss_aggrci}.}

\item{transform}{One of \code{"none"}, \code{"log"}, \code{"fisher"}, and \verb{"arcsin}.
Defaults to \code{"none"}.}

\item{conf_level}{Confidence level. Defaults to \code{0.95}.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{bootstrap}{If \code{TRUE}, performs a studentized bootstrap with \code{n_reps}
repetitions. Defaults to \code{FALSE}.}

\item{n_reps}{Number of bootstrap samples if \code{bootstrap = TRUE}. Ignored if
\code{bootstrap = FALSE}. Defaults to \code{1000}.}

\item{values}{to attach to each column on the Fleiss form data.
Defaults to \code{1:C}, where \code{C} is the number of categories. Only used
in \code{fleiss_aggr} and \code{bp_aggr}.}

\item{kind}{The kind of Brennan-Prediger coefficient used, \code{1} for the
classical kind and \code{2} for the kind introduced in Moss (2023). Only
relevant for \code{bp_aggr} and \code{bp}.}
}
\value{
A vector of class \code{quadagree} containing the confidence end points.
The arguments of the function call are included as attributes.
}
\description{
Calculate confidence intervals quadratically weighted multi-rater kappas
with potentially missing data. \code{congerci} and \code{cohenci} are aliases.
See \code{\link[=fleiss]{fleiss()}} and \code{\link[=cohen]{cohen()}} for estimation without inference.
}
\details{
The confidence intervals are based on the formulas of Moss and van Oest
(work in progress) along with standard asymptotic theory
(Magnus, Neudecker, 2019) and the missing data theory of
van Praag et al. (1985).

The methods handle missing data using pairwise available information, i.e.,
the option \code{use = "pairwise.complete.obs"} in \code{\link[stats:cor]{stats::cov()}} along with
the asymptotic theory of van Praag et al. (1985). The bootstrap option
uses the studentized bootstrap (Efron, B. 1987), which is second order
correct. Both functions makes use of \code{\link{future.apply}} when bootstrapping.

The \code{type} variables defaults to \code{adf}, asymptotically distribution-free,
which is consistent when the fourth moment is finite. The \code{normal} option
assumes normality, and is not consistent for models with excess
kurtosis unequal to \code{0}. The \code{elliptical} option assumes an
elliptical or pseudo-elliptical distribution of the data. The resulting
confidence intervals are corrected variants of the normal theory
intervals with a kurtosis correction (Yuan & Bentler 2002). The
common kurtosis parameter is calculated using the unbiased sample
kurtosis (Joanes, 1998).

Conger's (1980) kappa is a multi-rater generalization of Cohen's kappa.
All functions in this package work for multiple raters, so functions
starting with \code{cohen} or \code{conger} are aliases. The quadratically
weighted Cohen's kappa is also known as Lin's concordance coefficient.

The only difference between Cohen's kappa and Fleiss' kappa lies on how they
measure disagreement due to chance. Here Fleiss' marginalizes the rating
distribution across raters, essentially assuming there is no difference in
the rating distribution across raters, while Cohen's kappa does not.
There is a large literature comparing Fleiss' kappa to Cohen's kappa, and
there is no consensus on which to prefer.
}
\references{
Efron, B. (1987). Better Bootstrap Confidence Intervals. Journal of the
American Statistical Association, 82(397), 171-185.
https://doi.org/10.2307/2289144

Van Praag, B. M. S., Dijkstra, T. K., & Van Velzen, J. (1985).
Least-squares theory based on general distributional assumptions with
an application to the incomplete observations problem.
Psychometrika, 50(1), 25–36. https://doi.org/10.1007/BF02294145

Joanes, D. N., & Gill, C. A. (1998). Comparing measures of sample skewness
and kurtosis. Journal of the Royal Statistical Society: Series D
(The Statistician), 47(1), 183-189. https://doi.org/10.1111/1467-9884.00122

Cohen, J. (1968). Weighted kappa: Nominal scale agreement with provision for
scaled disagreement or partial credit. Psychological Bulletin, 70(4),
213–220. https://doi.org/10.1037/h0026256

Fleiss, J. L. (1975). Measuring agreement between two judges on the presence
or absence of a trait. Biometrics, 31(3), 651–659.
https://www.ncbi.nlm.nih.gov/pubmed/1174623

Conger, A. J. (1980). Integration and generalization of kappas for multiple
raters. Psychological Bulletin, 88(2), 322–328.
https://doi.org/10.1037/0033-2909.88.2.322

Lin, L. I. (1989). A concordance correlation coefficient to evaluate
reproducibility. Biometrics, 45(1), 255–268.
https://www.ncbi.nlm.nih.gov/pubmed/2720055

Moss, van Oest (work in progress). Inference for quadratically weighted
multi-rater kappas with missing raters.

Moss (work in progress). On the Brennan–Prediger coefficients.

Magnus, J. R., & Neudecker, H. (2019). Matrix Differential Calculus with
Applications in Statistics and Econometrics. John Wiley & Sons.
https://doi.org/10.1002/9781119541219
}
