% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{studentized_boots}
\alias{studentized_boots}
\title{Studentized bootstrap estimates using transformers.}
\usage{
studentized_boots(
  n_reps,
  x,
  type,
  transformer,
  fleiss = FALSE,
  kind = NULL,
  values = NULL
)
}
\arguments{
\item{n_reps}{Number of bootstrap repetitions.}

\item{x}{Data to estimate kappa on.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{transformer}{A \code{transformer} object.}

\item{fleiss}{If \code{TRUE}, calculates Fleiss' kappa. If not, calculates
Conger's kappa.}
}
\value{
Studentized bootstrap estimates.
}
\description{
Studentized bootstrap estimates using transformers.
}
\keyword{internal}
