% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci_asymptotic}
\alias{ci_boot}
\title{Calculates asymptotic confidence intervals.}
\usage{
ci_asymptotic(est, sd, n, transformer, quants)

ci_boot(x, est, sd, type, transformer, quants, n_reps, fleiss = FALSE)
}
\arguments{
\item{est, sd}{The estimate and estimated standard deviation.}

\item{n}{Number of observations.}

\item{transformer}{A transformer object.}

\item{quants}{Quantiles for the confidence interval.}

\item{x}{Data to estimate kappa on.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{n_reps}{Number of bootstrap samples if \code{bootstrap = TRUE}. Ignored if
\code{bootstrap = FALSE}.}

\item{fleiss}{If \code{TRUE}, calculates Fleiss' kappa. If not, calculates
Conger's kappa.}
}
\description{
Calculates asymptotic confidence intervals.
}
\keyword{internal}
