% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{cov_mat_kappa}
\alias{cov_mat_kappa}
\title{Covariance matrix (x,y,z) for Fleiss' and Cohen's kappa.}
\usage{
cov_mat_kappa(p, mu, sigma, gamma, x)
}
\arguments{
\item{p}{Vector of probabilities for being missing.}

\item{mu}{Vector of means.}

\item{sigma}{Covariance matrix}

\item{gamma}{Covariance matrix of the covariances.}
}
\value{
The covariance matrix of x, y, z
}
\description{
Returns the covariance matrix of x, y, z as defined in the paper on
quadratic Fleiss kappa and Cohen's kappa. These are defined as
$x=1^{T}\Sigma1-\tr\Sigma$, $y=\tr\Sigma$, and
$z=\left(\overline{\mu^{2}}-\overline{\mu}^{2}\right)$.
}
\keyword{internal}
