% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{fleiss}
\alias{fleiss}
\alias{conger}
\alias{cohen}
\alias{conger_pop}
\alias{cohen_pop}
\alias{fleiss_pop}
\title{Calculate the quadratically weighted Fleiss's kappa, Conger's kappa, and
Cohen's kappa}
\usage{
fleiss(x)

conger(x)

cohen(x)

conger_pop(mu, sigma)

cohen_pop(mu, sigma)

fleiss_pop(mu, sigma)
}
\arguments{
\item{x}{Input data on wide. Columns are judges, items are ratings. Missing
values are allowed.}

\item{mu}{Vector of means.}

\item{sigma}{Covariance matrix; must have the same dimensions as mu.}
}
\value{
Sample quadratically weighted Fleiss' kappa.

Population value of Conger's kappa / Fleiss' kappa.
}
\description{
Calculate the quadratically weighted Fleiss's kappa, Conger's kappa, and
Cohen's kappa

Calculates the population value of Conger's kappa and Fleiss' kappa.
}
\examples{
x <- irrCAC::cac.raw4raters[2:9, ]
fleiss(x)
# [1] 0.6666667
irrCAC::fleiss.kappa.raw(x, weights = "quadratic")$est$coeff.val
# [1] 0.66667 (irrCAC prematurely rounds the coefficient)

x <- irrCAC::cac.raw4raters[2:9, ]
conger(x)
# [1] 0.6719243
irrCAC::conger.kappa.raw(x, weights = "quadratic")$est$coeff.val
# [1] 0.67192 (irrCAC prematurely rounds the coefficient)
}
