% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_kappa.R
\name{get_diag_indices}
\alias{get_diag_indices}
\title{Get the indices of the diagonal for the variances.}
\usage{
get_diag_indices(r, vech = TRUE)
}
\arguments{
\item{r}{Number of raters.}

\item{vech}{If \code{TRUE}, returns for the result for half-vectorization.}
}
\value{
A vector of \code{TRUE} where the variances are.
}
\description{
Get the indices of the diagonal for the variances.
}
\keyword{internal}
