% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadagree.R
\docType{package}
\name{quadagree-package}
\alias{quadagree}
\alias{quadagree-package}
\title{quadagree: Inference for the quadratically weighted Fleiss' kappa, Conger's kappa, Cohen's kappa, and Brennan-Prediger coefficient}
\description{
Package for quadratically weighted agreement coefficients. See \verb{[quadagree]}
for basic documentation and browse the vignettes for more details.
\code{browseVignettes(package = "quadagree")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonasmoss.github.io/quadagree/}
}

}
\author{
\strong{Maintainer}: Jonas Moss \email{jonas.moss.statistics@gmail.com} (\href{https://orcid.org/0000-0002-6876-6964}{ORCID})

}
