% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{studentized_boots_aggr}
\alias{studentized_boots_aggr}
\title{Studentized bootstrap estimates using transformers.}
\usage{
studentized_boots_aggr(n_reps, x, values, transformer)
}
\arguments{
\item{n_reps}{Number of bootstrap repetitions.}

\item{x}{Data to estimate kappa on.}

\item{values}{Values for the different categories.}

\item{transformer}{A \code{transformer} object.}

\item{fleiss}{If \code{TRUE}, calculates Fleiss' kappa. If not, calculates
Conger's kappa.}
}
\value{
Studentized bootstrap estimates.
}
\description{
Studentized bootstrap estimates using transformers.
}
\keyword{internal}
